requirements.txt
openalgo
gspread
oauth2client
gspread-formatting
python-dotenv